+++ sat/cli/nid2xname/parser.py
@@ -1,5 +1,6 @@
 """
 The parser for the nid2xname subcommand.
+
 (C) Copyright 2021 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a