+++ sat/cli/xname2nid/parser.py
@@ -1,5 +1,6 @@
 """
 The parser for the xname2nid subcommand.
+
 (C) Copyright 2021 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a