+++ tests/cli/xname2nid/test_xname2nid.py
@@ -1,7 +1,7 @@
 """
-Unit tests for the sat.cli.xname2nid
+Unit tests for the sat.cli.xname2nid module.
 
-(C) Copyright 2020 Hewlett Packard Enterprise Development LP.
+(C) Copyright 2021 Hewlett Packard Enterprise Development LP.
 
 Permission is hereby granted, free of charge, to any person obtaining a
 copy of this software and associated documentation files (the "Software"),
@@ -22,11 +22,14 @@ ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 OTHER DEALINGS IN THE SOFTWARE.
 """
 
+import logging
 import unittest
-from unittest import mock
 from argparse import Namespace
+from unittest import mock
 
+from sat.apiclient import APIError
 from sat.cli.xname2nid.main import do_xname2nid
+from tests.common import ExtendedTestCase
 
 
 def set_options(namespace):
@@ -34,14 +37,12 @@ def set_options(namespace):
     namespace.xnames = ['x1000c0s1b0n1']
 
 
-class TestDoXname2nid(unittest.TestCase):
+class TestDoXname2nid(ExtendedTestCase):
     """Unit test for xname2nid"""
 
     def setUp(self):
         """Mock functions called."""
-        self.mock_hsm_client = mock.patch('sat.cli.xname2nid.main.HSMClient',
-                                          autospec=True).start().return_value
-        self.mock_hsm_client.get_node_components.return_value = [
+        self.node_data = [
             {'NID': 1006,
              'Type': 'Node',
              'ID': 'x1000c0s1b0n1'},
@@ -55,6 +56,10 @@ class TestDoXname2nid(unittest.TestCase):
              'Type': 'Node',
              'ID': 'x1000c2s2b0n1'}
         ]
+        self.mock_hsm_client = mock.patch('sat.cli.xname2nid.main.HSMClient',
+                                          autospec=True).start().return_value
+        self.mock_hsm_client.get_node_components.return_value = self.node_data
+
         self.mock_sat_session = mock.patch('sat.cli.xname2nid.main.SATSession').start()
         self.mock_print = mock.patch('builtins.print', autospec=True).start()
 
@@ -67,38 +72,72 @@ class TestDoXname2nid(unittest.TestCase):
 
     def test_one_xname_exists(self):
         """Test do_xname2nid with one valid xname."""
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        self.assert_in_element(f'xname: {self.node_data[0]["ID"]}, nid: {self.node_data[0]["NID"]}',
+                               logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('nid001006')
 
     def test_one_xname_not_exists(self):
         """Test do_xname2nid with one invalid xname."""
         self.fake_args.xnames = ['x1000c2s1b0n5']
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        self.assert_in_element(f'xname: {self.fake_args.xnames[0]}, nid: MISSING', logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('MISSING')
 
     def test_two_xnames_exist(self):
         """Test do_xname2nid with two valid xnames."""
         self.fake_args.xnames = ['x1000c2s1b0n0', 'x1000c2s2b0n0']
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        for node in self.node_data[1:2]:
+            self.assert_in_element(f'xname: {node["ID"]}, nid: {node["NID"]}', logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('nid001069,nid001073')
 
     def test_two_comma_separated_xnames_exist(self):
         """Test do_xname2nid with two valid comma separated xnames."""
         self.fake_args.xnames = ['x1000c2s1b0n0,x1000c2s2b0n0']
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        for node in self.node_data[1:2]:
+            self.assert_in_element(f'xname: {node["ID"]}, nid: {node["NID"]}', logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('nid001069,nid001073')
 
     def test_two_comma_separated_xnames_with_space_exist(self):
         """Test do_xname2nid with two valid comma separated xnames with space."""
         self.fake_args.xnames = ['x1000c2s1b0n0,', 'x1000c2s2b0n0']
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        for node in self.node_data[1:2]:
+            self.assert_in_element(f'xname: {node["ID"]}, nid: {node["NID"]}', logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('nid001069,nid001073')
 
     def test_three_xnames_one_invalid(self):
         """Test do_xname2nid with two valid xnames and one invalid string."""
         self.fake_args.xnames = ['x1000c2s1b0n0', 'not-an-xname', 'x1000c2s2b0n0']
-        do_xname2nid(self.fake_args)
+        with self.assertLogs(level=logging.INFO) as logs:
+            do_xname2nid(self.fake_args)
+        for node in self.node_data[1:2]:
+            self.assert_in_element(f'xname: {node["ID"]}, nid: {node["NID"]}', logs.output)
+        self.assert_in_element(f'xname: {self.fake_args.xnames[1]}, nid: MISSING', logs.output)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
         self.mock_print.assert_called_once_with('nid001069,MISSING,nid001073')
+
+    def test_xname2nid_api_error(self):
+        """Test xname2nid logs an error and exits when an APIError occurs."""
+        self.mock_hsm_client.get_node_components.side_effect = APIError('HSM failed')
+        with self.assertLogs(level=logging.ERROR) as logs:
+            with self.assertRaises(SystemExit):
+                do_xname2nid(self.fake_args)
+        self.mock_hsm_client.get_node_components.assert_called_once_with()
+        self.assert_in_element('Request to HSM API failed: HSM failed', logs.output)
+        self.mock_print.assert_not_called()
 
 
 if __name__ == '__main__':