+++ CHANGELOG.md
@@ -35,6 +35,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ``uan_bos_template``.
 - Added a confirmation message when generating a configuration file via
   ``sat init``.
+- Added a new ``sat nid2xname`` subcommand to translate node IDs to node xnames.
+- Added a new ``sat xname2nid`` subcommand to translate node and node BMC xnames
+  to node IDs.
 
 ### Changed
 - ``--cle-bos-template`` and ``--uan-bos-template`` now no longer have defaults